fixed some of the bugs, including the newlist component staying open when newCard is open
