--- productab/src/redux/reducers/root.js
@@ -1,8 +1,0 @@
-import { combineReducers } from "redux";
-import { base } from "./base";
-
-const rootReducer = combineReducers({
-  base
-});
-
-export default rootReducer;
