+++ productab_client/public/manifest.json
