+++ productab_client/src/App.js
@@ -2,6 +2,8 @@ import React, { Component } from "react";
 import Container from './containers/container';
 import { connect } from "react-redux";
 
+// import *as _ from "./redux/actions/baseActions";
+import "./_stylesheet/index.scss";
 import "./_stylesheet/App.scss";
 import "./_stylesheet/components.scss";
 import "./_stylesheet/forms.scss";
