+++ productab_client/src/_stylesheet/App.scss
@@ -3,16 +3,11 @@
 div.App {
   height: 100vh;
   min-width: 100vw;
-  overflow-x: scroll;
 }
 
 div#container {
   display: grid;
   height: 100vh;
-  grid-template: 
-    "head" 10vh
-    "main" auto
-    "side" 10vh / 100vw;
 }
 
 div#top-panel {
@@ -29,20 +24,23 @@ div#side-panel {
 
 div#main-content {
   grid-area: main;
+  background-color: $default-bg;
+  box-sizing: border-box;
+}
+
+div#inner-content {
+  width: fit-content;
   padding: $main-padding;
   padding-left: 0px;
   display: flex;
   flex-direction: row;
-  overflow-x: scroll;
-  background-color: $default-bg;
-  box-sizing: border-box;
 }
 
 @media (min-width: $breakpoint-tablet-landscape) {
   div#container{
     grid-template:
       "side head" 10vh
-      "side main" 1fr / 1fr 95vw;
+      "side main" 1fr / 1fr 90vw;
   }
 }
