+++ productab_client/src/_stylesheet/components.scss
@@ -1,7 +1,9 @@
 @import "index";
 
 .shrink-container {
-  display: table;
+  display: flex;
+  flex-direction: column;
+  height: fit-content;
 }
 
 .transparent-black {
@@ -10,13 +12,13 @@
 
 div#main-content {
   overflow-x: scroll;
+  // width: fit-content;
 }
 
 div.list-sizing {
   max-width: $list-width;
   width: $list-width;
   overflow-x: hidden;
-  // width: fit-content;
   box-sizing: border-box;
   margin-left: $main-padding;
   border-radius: $border-radius;
@@ -27,6 +29,7 @@ div.list-style{
   box-shadow: 5px 5px 8px 0px rgba($dark-grey, 0.5);
 
   div.list-top {
+    margin-bottom: $main-padding;
     min-height: $list-cap-height;
     display: flex;
     flex-direction: row;
@@ -48,7 +51,7 @@ div.list-style{
 
   div.list-bottom {
     min-height: $list-cap-height;
-    padding: 0 $main-padding;
+    padding: $main-padding $main-padding 0;
   }
 }
 
@@ -80,3 +83,52 @@ i.fa-ellipsis-h {
 i.fa-ellipsis-h:hover {
   cursor: pointer;
 }
+
+div.add-card-div:hover {
+  background-color: rgba($dark-grey, 0.7);
+  color: white;
+  cursor: pointer;
+}
+
+div.inner-list {
+  display: flex;
+  flex-direction: column;
+  box-sizing: content-box;
+  padding: $main-padding/2 $main-padding;
+  // padding-bottom: $main-padding/2;
+}
+
+div.card {  
+  margin: $main-padding/2 0;
+  display: grid;
+  grid-template: 
+    "title edit" auto/ auto $main-padding*3;
+  box-sizing: border-box;
+  width: auto;
+  max-height: 120px;
+  // margin-bottom: $main-padding/2;
+  // padding: $main-padding/4 $main-padding;
+  border-radius: $border-radius;
+  background-color: white;
+
+
+}
+div.card:hover {  
+  cursor: pointer;
+  background-color: rgba($dark-grey, 0.3);
+}
+
+
+
+div.card-settings {
+  margin: auto;
+  grid-area: edit;
+}
+
+  div.card-title {
+    grid-area: title;
+
+    p.card-title-contents  {
+      margin-left: $main-padding;
+    }
+  }
