+++ productab_client/src/_stylesheet/index.scss
@@ -5,11 +5,12 @@ $list-spacer: 3vw;
 $main-padding: 10px;
 $list-cap-height: 30px;
 $list-width: 270px;
+$inner-width: inherit - 2*$main-padding;
 
 /* Other Variables */
 $test-border: 3px solid;
 $default-grey: #dbe0e0;
-$dark-grey: #353535;
+$dark-grey: #1d1d1d;
 $default-bg: #429c6b;
 $dark-green: #217547;
 $lt-blue: #2d55aa;
@@ -37,7 +38,7 @@ textarea {
   border: none;
   border-style: none;
   font-size: 1em;
-  max-width: inherit - 2*$main-padding;
+  max-width: $inner-width;
 }
 
 textarea:focus {
