+++ productab_client/src/containers/mainContent.js
@@ -3,7 +3,9 @@ import { connect } from "react-redux";
 
 import List from "./list";
 import NewList from "../components/newList";
-import * as _ from "../redux/actions/baseActions";
+import * as _base from "../redux/actions/baseActions";
+import * as _list from "../redux/actions/listActions";
+import * as _card from "../redux/actions/cardActions";
 import "../_stylesheet/components.scss";
 
 class MainContent extends React.Component {
@@ -27,15 +29,19 @@ class MainContent extends React.Component {
   }
 
   handleBlurClick = (evt) => {
-    const { disableForms } = this.props;
-    if (evt.target.id === "main-content")
+    const { disableForms, deselectList } = this.props;
+    const targetId = evt.target.id;
+
+    if (targetId === "main-content" || targetId === "inner-content")
     {
       disableForms();
+      deselectList();
     }
     evt.persist();
+    evt.preventDefault();
   }
 
-  /* Wraps array of current board's lists into React Components to be rendered */
+  /* Returns array of current board's lists as React Components to be rendered */
   showLists = () => {
     const { lists } = this.props;
     return lists.map(list => <List key={`Board. ${list.board_id} - No. ${list.id}`} list={list}/>);
@@ -43,22 +49,24 @@ class MainContent extends React.Component {
 
   render() {
     return(<div id="main-content" onClick={this.handleBlurClick}>
-      {/* <List/> */}
-      {this.showLists()}
-      <NewList/>
+      <div id="inner-content">
+        {this.showLists()}
+        <NewList/>
+      </div>
     </div>);
   }
 }
 
 const mapStateToProps = (state) => ({
-  lists: state.base.lists,
-  cards: state.base.cards
+  lists: state.list.lists,
+  cards: state.card.cards
 });
 
 const mapDispatchToProps = (dispatch) => ({
-  disableForms: () => dispatch(_.disableForms()),
-  getLists: (boardId) => dispatch(_.getLists(boardId)),
-  getCards: () => dispatch(_.getCards())
+  disableForms: () => dispatch(_base.disableForms()),
+  getLists: (boardId) => dispatch(_list.getLists(boardId)),
+  getCards: () => dispatch(_card.getCards()),
+  deselectList: () => dispatch(_list.deselectList())
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(MainContent);
