+++ productab_client/src/containers/sidePanel.js
