+++ productab_client/src/scratch.js
@@ -1,0 +1,1 @@
+// Basically a code scratchpad used to refactor code and keep an alpha and beta version just in case of SNAFUs
